k: Move partition_proxy to nascent kafka/data directory
Home for kafka data plane oriented utils and interfaces.
