+++ src/v/cloud_topics/reconciler/reconciler.cc
@@ -12,7 +12,7 @@
 
 #include "base/vlog.h"
 #include "cluster/partition.h"
-#include "kafka/server/partition_proxy.h"
+#include "kafka/data/partition_proxy.h"
 #include "kafka/utils/txn_reader.h"
 #include "model/namespace.h"
 #include "random/generators.h"
