+++ src/v/cluster/BUILD
@@ -200,9 +200,9 @@ redpanda_cc_library(
         "//src/v/cluster",
         "//src/v/config",
         "//src/v/container:fragmented_vector",
+        "//src/v/kafka/data:partition_proxy",
         "//src/v/kafka/protocol",
         "//src/v/kafka/server",
-        "//src/v/kafka/server:partition_proxy",
         "//src/v/model",
         "//src/v/rpc",
         "//src/v/serde",
