+++ src/v/cluster/cloud_metadata/offsets_lookup.cc
@@ -16,7 +16,7 @@
 #include "cluster/partition_manager.h"
 #include "cluster/shard_table.h"
 #include "config/configuration.h"
-#include "kafka/server/partition_proxy.h"
+#include "kafka/data/partition_proxy.h"
 #include "model/ktp.h"
 
 #include <seastar/util/later.hh>
