+++ src/v/kafka/data/BUILD
@@ -1,0 +1,31 @@
+load("//bazel:build.bzl", "redpanda_cc_library")
+
+redpanda_cc_library(
+    name = "partition_proxy",
+    srcs = [
+        "partition_proxy.cc",
+        "replicated_partition.cc",
+    ],
+    hdrs = [
+        "partition_proxy.h",
+        "replicated_partition.h",
+    ],
+    implementation_deps = [
+        "//src/v/cluster",
+        "@boost//:numeric_conversion",
+    ],
+    include_prefix = "kafka/data",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//src/v/base",
+        "//src/v/cloud_storage",
+        "//src/v/kafka/protocol",
+        "//src/v/kafka/server:errors",
+        "//src/v/model",
+        "//src/v/raft",
+        "//src/v/storage",
+        "//src/v/utils:truncating_logger",
+        "@boost//:outcome",
+        "@seastar",
+    ],
+)
