+++ src/v/kafka/data/partition_proxy.h
