+++ src/v/kafka/data/replicated_partition.h
@@ -11,8 +11,8 @@
 #pragma once
 
 #include "cloud_storage/types.h"
+#include "kafka/data/partition_proxy.h"
 #include "kafka/protocol/errors.h"
-#include "kafka/server/partition_proxy.h"
 #include "model/fundamental.h"
 #include "model/record_batch_reader.h"
 #include "raft/replicate.h"
