+++ src/v/kafka/server/handlers/fetch.cc
@@ -15,6 +15,8 @@
 #include "cluster/shard_table.h"
 #include "config/configuration.h"
 #include "container/fragmented_vector.h"
+#include "kafka/data/partition_proxy.h"
+#include "kafka/data/replicated_partition.h"
 #include "kafka/protocol/batch_consumer.h"
 #include "kafka/protocol/errors.h"
 #include "kafka/protocol/fetch.h"
@@ -25,9 +27,7 @@
 #include "kafka/server/handlers/fetch/fetch_planner.h"
 #include "kafka/server/handlers/fetch/replica_selector.h"
 #include "kafka/server/latency_probe.h"
-#include "kafka/server/partition_proxy.h"
 #include "kafka/server/read_distribution_probe.h"
-#include "kafka/server/replicated_partition.h"
 #include "model/fundamental.h"
 #include "model/metadata.h"
 #include "model/namespace.h"
