+++ src/v/kafka/server/handlers/list_offsets.cc
@@ -13,11 +13,11 @@
 #include "cluster/partition_manager.h"
 #include "cluster/shard_table.h"
 #include "container/fragmented_vector.h"
+#include "kafka/data/partition_proxy.h"
+#include "kafka/data/replicated_partition.h"
 #include "kafka/protocol/errors.h"
 #include "kafka/server/errors.h"
 #include "kafka/server/handlers/details/leader_epoch.h"
-#include "kafka/server/partition_proxy.h"
-#include "kafka/server/replicated_partition.h"
 #include "kafka/server/request_context.h"
 #include "kafka/server/response.h"
 #include "model/fundamental.h"
