+++ src/v/kafka/server/handlers/offset_for_leader_epoch.cc
@@ -12,10 +12,10 @@
 #include "cluster/metadata_cache.h"
 #include "cluster/shard_table.h"
 #include "container/fragmented_vector.h"
+#include "kafka/data/partition_proxy.h"
 #include "kafka/protocol/errors.h"
 #include "kafka/protocol/schemata/offset_for_leader_epoch_response.h"
 #include "kafka/server/handlers/details/leader_epoch.h"
-#include "kafka/server/partition_proxy.h"
 #include "kafka/server/request_context.h"
 #include "model/fundamental.h"
 #include "model/namespace.h"
