+++ src/v/kafka/utils/txn_reader.cc
@@ -11,7 +11,7 @@
 
 #include "txn_reader.h"
 
-#include "kafka/server/partition_proxy.h"
+#include "kafka/data/partition_proxy.h"
 #include "model/fundamental.h"
 #include "model/record.h"
 #include "model/record_batch_reader.h"
