+++ src/v/transform/rpc/BUILD
@@ -31,7 +31,7 @@ redpanda_cc_library(
         "//src/v/cluster",
         "//src/v/config",
         "//src/v/container:fragmented_vector",
-        "//src/v/kafka/server:partition_proxy",
+        "//src/v/kafka/data:partition_proxy",
         "//src/v/model",
         "//src/v/raft",
         "//src/v/resource_mgmt:io_priority",
