+++ src/v/transform/rpc/deps.h
@@ -12,7 +12,7 @@
 
 #include "cluster/fwd.h"
 #include "cluster/types.h"
-#include "kafka/server/partition_proxy.h"
+#include "kafka/data/partition_proxy.h"
 #include "model/fundamental.h"
 #include "model/ktp.h"
 #include "model/metadata.h"
